HIVE-29750: Iceberg: Move Iceberg REST catalog client code to a new module
