+++ itests/hive-iceberg/pom.xml
@@ -64,7 +64,7 @@
     </dependency>
     <dependency>
       <groupId>org.apache.hive</groupId>
-      <artifactId>hive-iceberg-catalog</artifactId>
+      <artifactId>hive-iceberg-rest-catalog-client</artifactId>
       <version>${project.version}</version>
     </dependency>
     <dependency>
