+++ packaging/src/docker/thirdparties/gravitino/hive/hive-site.xml
@@ -90,7 +90,7 @@
 
   <property>
     <name>metastore.client.impl</name>
-    <value>org.apache.iceberg.hive.client.HiveRESTCatalogClient</value>
+    <value>org.apache.iceberg.hive.rest.catalog.client.HiveRESTCatalogClient</value>
   </property>
 
   <property>
